tests of idempotent upsert
